backend/requirements.txt
Django==4.2.23
djangorestframework==3.14.0
djoser==2.2.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
dj-database-url==2.1.0
Pillow==11.2.1
django-cors-headers==4.7.0 
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
